requirements.txt
Flask
Flask-Bcrypt
boto3
python-dotenv
Flask-Cors
PyJWT
openai
